fix: resolve compilation issues after folders moving
