+++ src/components/Header/components/Avatar/index.ts
@@ -1,0 +1,3 @@
+import UserAvatar from "./Avatar";
+
+export default UserAvatar;
