+++ src/pages/HomePage.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import { Suspense } from 'react';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import { SearchContext } from '../context/SearchContext';
-import Intro from '../components/intro';
-import CountryCardsContainer from '../components/country-cards-container';
+import Intro from '../components/Intro';
+import CountryCardsContainer from '../components/CountryCardsContainer';
 
 export const HomePage: React.FC = () => {
   const [searchText, setSearchText] = React.useState('');
